Update audioaip
This changes the process slightly so that if more than one WAV file is present for a tape (ex. Side A and Side B) that both are packaged as part of a single AIP for each audio tape. I also changed the data integrity verification from manifest-md5 to manifest-sha256 to improve security.